SDK V2 Updated Changes (According to PR Comments)
- Updated documentation link to AWS SDK v2
- Changed context.TODO() to context.Background() for initialization
- Made StorageCapacity int32 throughout codebase
- Made newSizeGiB parameter int32 for resize operations
- Used ErrorMessage() method instead of direct field access
- Updated AWS SDK packages to latest versions (FSx v1.55.0)
- Updated Go version to 1.24.0
- Removed unnecessary log statement for consistency
- Changed getter functions to return by value instead of pointer
- Refactored getTagsForResource to accept context parameter
- Added client-side validation for delete parameters
- Replaced fsxtypes alias with types for consistency
- Restored EC2 metadata availability check using GetInstanceIdentityDocument
- Added back missing test functions updated for AWS SDK v2
- Fixed all compilation errors and test failures

+++ pkg/driver/controller_test.go
@@ -38,7 +38,7 @@ func TestCreateVolume(t *testing.T) {
 		snapshotVolumeParameters     map[string]string
 		filesystemId                       = "filesystemId"
 		volumeId                           = "volumeId"
-		storageCapacity              int64 = 64
+		storageCapacity              int32 = 64
 		dnsName                            = "dnsName"
 		snapshotId                         = "fsvolsnap-1234"
 		volumePath                         = "/"
@@ -1698,9 +1698,9 @@ func TestControllerExpandVolume(t *testing.T) {
 		dnsName               = "dnsName"
 		filesystemId          = "fs-1234"
 		volumeId              = "fsvol-1234"
-		storageCapacity int64 = 100
+		storageCapacity int32 = 100
 		currentBytes          = util.GiBToBytes(storageCapacity)
-		newCapacity     int64 = 150
+		newCapacity     int32 = 150
 		requiredBytes         = util.GiBToBytes(newCapacity)
 		limitBytes            = util.GiBToBytes(newCapacity)
 	)
